requirements.txt
# Core dependencies
pyserial>=3.5
SpeechRecognition>=3.8.1
pyttsx3>=2.90

# Modern TTS
gTTS>=2.3.1
pygame>=2.1.2
edge-tts>=6.1.9

# LLM integration
subprocess.run

# Optional: Coqui TTS (uncomment if needed)
# TTS>=0.13.0

# Optional: Piper TTS (install system-wide)
# sudo apt install piper-tts
